+++ maqs-accessibility/src/main/java/com/cognizantsoftvision/maqs/accessibility/AccessibilityUtilities.java
@@ -257,7 +257,7 @@ public class AccessibilityUtilities {
       }
 
       results = getResults.get();
-      HtmlReporter.createAxeHtmlReport(testObject.getWebDriver(), results, report, requestedResults);
+      SeleniumReporter.createAxeHtmlReport(testObject.getWebDriver(), results, report, requestedResults);
     } finally {
       // Restore logging if we suspended it
       if (restoreLogging) {
